+++ Alarm/Controllers/View Controllers/SwitchTableViewCell.swift
@@ -1,0 +1,36 @@
+//
+//  SwitchTableViewCell.swift
+//  Alarm
+//
+//  Created by theevo on 3/2/20.
+//  Copyright © 2020 theevo. All rights reserved.
+//
+
+import UIKit
+
+class SwitchTableViewCell: UITableViewCell {
+
+    // MARK: - IBOutlets
+    @IBOutlet weak var timeLabel: UILabel!
+    @IBOutlet weak var nameLabel: UILabel!
+    @IBOutlet weak var alarmSwitch: UISwitch!
+    
+    // MARK: - Lifecycle functions
+    
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+    
+    // MARK: - IBActions
+
+    @IBAction func switchValueChanged(_ sender: Any) {
+    }
+    
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
