--- myAlarm/Controllers/View Controllers/SwitchTableViewCell.swift
@@ -1,23 +1,0 @@
-//
-//  SwitchTableViewCell.swift
-//  myAlarm
-//
-//  Created by Jake Haslam on 3/2/20.
-//  Copyright © 2020 trevorAdcock. All rights reserved.
-//
-
-import UIKit
-
-class SwitchTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-    }
-
-}
